+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 * alphaSetOrgShortName(id: ID! @organizationID, shortname: String!): String
+* Organization.shortname: String